SoundBrowser: Folder entry bug fix
The folder was being checked for existence during entry of a new folder record, causing problems with lost input. From now on, the folder form only checks for existence in view mode (which is when the populate and open buttons should only be enabled if the folder exists).
Also, RMI file support was added to the sound file record's browse dialog (which lets you select a sound file name from your computer to fill the form).